Add endpoint to get dormitory by ID. Change ErrorDto to be record and create utility method in GlobalExceptionHandler for creating ErroDto

+++ src/main/java/com/dormify/dormitories/DormitoryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DormitoryRepository extends JpaRepository<Dormitory, Integer> {
+public interface DormitoryRepository extends JpaRepository<Dormitory, Long> {
 }
